Split index.tsx logic into components and generic welcome texts
